refactor: split path file
